Fixed commands

+++ eco-core/core-plugin/src/main/kotlin/com/willfp/ecobits/commands/CommandGive.kt
@@ -63,7 +63,7 @@ class CommandGive(
             return
         }
 
-        player.adjustBalance(currency, amount)
+        player.adjustBalance(currency, amount.toBigDecimal())
 
         sender.sendMessage(
             plugin.langYml.getMessage("gave-currency", StringUtils.FormatOption.WITHOUT_PLACEHOLDERS)
