+++ eco-core/core-plugin/src/main/kotlin/com/willfp/ecobits/commands/CommandTake.kt
@@ -64,7 +64,7 @@ class CommandTake(
             return
         }
 
-        player.adjustBalance(currency, -amount)
+        player.adjustBalance(currency, -amount.toBigDecimal())
 
         sender.sendMessage(
             plugin.langYml.getMessage("took-currency", StringUtils.FormatOption.WITHOUT_PLACEHOLDERS)
